bot/requirements.txt
nonebot2[fastapi,websockets]>=2.2.0
nonebot-adapter-onebot>=2.4.0
openai
nonebot-plugin-apscheduler
httpx
feedparser>=6.0.0
trafilatura
readability-lxml
beautifulsoup4
